Fixed all the new warnings.

+++ devices/v775/Device.cxx
@@ -65,7 +65,7 @@ namespace v775 {
          std::unique_ptr< ChannelConfig > channel( new ChannelConfig() );
          CHECK( channel->readConfig( dev ) );
          if( ( channel->getChannel() >= 0 ) &&
-             ( channel->getChannel() < NUMBER_OF_CHANNELS ) ) {
+             ( channel->getChannel() < static_cast<int>(NUMBER_OF_CHANNELS) ) ) {
             if( m_channels[ channel->getChannel() ] ) {
                m_logger << msg::WARNING
                         << tr( "Redefining channel number: %1" )
@@ -101,7 +101,7 @@ namespace v775 {
 
       // Count the number of configured channels:
       quint32 number_of_channels = 0;
-      for( int i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
+      for( size_t i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
          if( m_channels[ i ] ) ++number_of_channels;
       }
 
@@ -109,7 +109,7 @@ namespace v775 {
       output << number_of_channels;
 
       // Write the channel configurations:
-      for( int i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
+      for( size_t i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
          if( m_channels[ i ] ) {
             CHECK( m_channels[ i ]->writeConfig( dev ) );
          }
@@ -188,7 +188,7 @@ namespace v775 {
 
          // Make sure that the reading was successful:
          if( ( channel->getChannel() >= 0 ) &&
-             ( channel->getChannel() < NUMBER_OF_CHANNELS ) ) {
+             ( channel->getChannel() < static_cast<int>(NUMBER_OF_CHANNELS) ) ) {
             if( m_channels[ channel->getChannel() ] ) {
                m_logger << msg::WARNING
                         << tr( "Redefining channel number: %1" )
@@ -224,7 +224,7 @@ namespace v775 {
       element.setAttribute( "FullScaleRange", m_fullScaleRangeValue );
 
       // Create a new node for the configuration of each channel:
-      for( int i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
+      for( size_t i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
          if( m_channels[ i ] ) {
             QDomElement ch_element =
                element.ownerDocument().createElement( "Channel" );
