+++ devices/v792/Device.cxx
@@ -51,7 +51,7 @@ namespace v792 {
          std::unique_ptr< ChannelConfig > channel( new ChannelConfig() );
          CHECK( channel->readConfig( dev ) );
          if( ( channel->getChannel() >= 0 ) &&
-             ( channel->getChannel() < NUMBER_OF_CHANNELS ) ) {
+             ( channel->getChannel() < static_cast<int>(NUMBER_OF_CHANNELS) ) ) {
             if( m_channels[ channel->getChannel() ] ) {
                m_logger << msg::WARNING
                         << tr( "Redefining channel number: %1" )
@@ -86,7 +86,7 @@ namespace v792 {
 
       // Count the number of configured channels:
       quint32 number_of_channels = 0;
-      for( int i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
+      for( size_t i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
          if( m_channels[ i ] ) ++number_of_channels;
       }
 
@@ -94,7 +94,7 @@ namespace v792 {
       output << number_of_channels;
 
       // Write the channel configurations:
-      for( int i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
+      for( size_t i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
          if( m_channels[ i ] ) {
             CHECK( m_channels[ i ]->writeConfig( dev ) );
          }
@@ -157,7 +157,7 @@ namespace v792 {
 
          // Make sure that the reading was successful:
          if( ( channel->getChannel() >= 0 ) &&
-             ( channel->getChannel() < NUMBER_OF_CHANNELS ) ) {
+             ( channel->getChannel() < static_cast<int>(NUMBER_OF_CHANNELS) ) ) {
             if( m_channels[ channel->getChannel() ] ) {
                m_logger << msg::WARNING
                         << tr( "Redefining channel number: %1" )
@@ -192,7 +192,7 @@ namespace v792 {
       element.setAttribute( "Pedestal", m_pedestal );
 
       // Create a new node for the configuration of each channel:
-      for( int i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
+      for( size_t i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
          if( m_channels[ i ] ) {
             QDomElement ch_element =
                element.ownerDocument().createElement( "Channel" );
